requirements.txt
# =====================================================
# Aboud Trading Bot - Dependencies
# =====================================================
# Python 3.10+

# Web Framework
flask==3.1.0
gunicorn==23.0.0

# Telegram Bot
python-telegram-bot==21.9

# HTTP Client (async)
aiohttp==3.11.11

# Scheduler
APScheduler==3.11.0

# Utilities
python-dotenv==1.0.1
